unit test withoug  global, namespaced filesytem classes
